Implementando outras exceçoes

+++ src/main/java/com/devsuperior/dscommerce/services/ProductService.java
@@ -3,11 +3,15 @@ package com.devsuperior.dscommerce.services;
 import com.devsuperior.dscommerce.dto.ProductDTO;
 import com.devsuperior.dscommerce.entities.Product;
 import com.devsuperior.dscommerce.repositories.ProductRepository;
+import com.devsuperior.dscommerce.services.exceptions.DatabaseException;
 import com.devsuperior.dscommerce.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
@@ -43,17 +47,32 @@ public class ProductService {
 
     @Transactional
     public ProductDTO update(Long id,ProductDTO dto){
+        try{
 
-        Product entity = repository.getReferenceById(id);
-        copyDtotoEntity(dto,entity);
-        entity = repository.save(entity);
-        return new ProductDTO(entity);
+            Product entity = repository.getReferenceById(id);
+            copyDtotoEntity(dto,entity);
+            entity = repository.save(entity);
+            return new ProductDTO(entity);
+
+        }
+        catch (EntityNotFoundException e){
+
+            throw new ResourceNotFoundException("Recurso nao encontrado");
+        }
+
     }
 
-    @Transactional
-    public void delete(Long id){
-          repository.deleteById(id);
-
+    @Transactional(propagation = Propagation.SUPPORTS)
+    public void delete(Long id) {
+        if (!repository.existsById(id)) {
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
+        try {
+            repository.deleteById(id);
+        }
+        catch (DataIntegrityViolationException e) {
+            throw new DatabaseException("Falha de integridade referencial");
+        }
     }
 
     private void copyDtotoEntity(ProductDTO dto, Product entity) {
